requirements.txt
openai==1.12.0
firebase-admin==6.4.0
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pandas==2.1.3
openpyxl==3.1.2
httpx==0.26.0
PyPDF2==3.0.1
requests==2.31.0 